Adds fonts folder & work-sans
Fonts are important to design!  This folder gives org members a place to go to get all open source / free fonts that we use.

Work Sans is licensed under the SIL Open Font License and as such they are cool with us redistributing it.  Licenses should be included in their respective font folder.